Add Discord webhook exposed key handling
